fix clang warnings

+++ source/newtonian/two_dimensional/ghost_point_generators/PeriodicGhostGenerator.cpp
@@ -25,5 +25,5 @@ std::pair<ComputationalCell, ComputationalCell> PeriodicGhostGenerator::GetGhost
 Vector2D PeriodicGhostGenerator::GetGhostVelocity(const Tessellation& tess, const vector<ComputationalCell>& /*cells*/,
 	vector<Vector2D> const& point_veolcities, size_t ghost_index, Edge const& /*edge*/)const
 {
-	return point_veolcities[tess.GetOriginalIndex(static_cast<int>(ghost_index))];
+  return point_veolcities[static_cast<size_t>(tess.GetOriginalIndex(static_cast<int>(ghost_index)))];
 }
